Remove etce.conf SSH configuration parameters. Use paramiko.client.RejectPolicy always when a destination host is not found in the users known_hosts file.
Change SSHClient to use ~/.ssh/config values for Port, User
and IdentityFile for creating ssh connections when these values
are not specified on the command line. Use port=22 and
user=os.path.expand('~') when Port and User are not specified
on the command line or in ~/.ssh/config.

Rename etce-exec.ssh to etce-field-exec.