Add optional decorator attribute to stepsfile run element. Currently the only permitted value is "sudo":
   <run wrapper="emane" decorator="sudo"/>

Specifying a decorator has the effect of passing the wrapper build
command through a second processing phase (the decorator) that
has the opportunity to modify the command. For the sudo decorator
this means prepending "sudo" to the command.

Remove the LOCK_FILE_DIRECTORY option from etce.conf. Fix the
lock directory to be WORK_DIRECTORY/lock for wrappers. For
etce-lxc, the lock file is saved to WORK_DIRECTORY/lxcroot
alongside the LXC configurations.