=== Add policy argument to SSHClient to set the policy to follow when a destination host is not present in the local known_hosts file:
   reject:   reject the host and quit
   warning:  warn that the host is not trusted but proceed
   autoadd:  automatically add the host to known_hosts.

The default policy remains reject.

===
Add policy command line arguments to scripts that act as an SSH client.

===
Add arguments to etce-kill:

   nosudo: do not preface the kill command with sudo on the remote hosts
   signal: specify the numeric signal to use to kill

===
Change etce-field-exec to create the WORKING_DIRECTORY when it is not present.

===
Fix a bug in the platform kill command that allowed a host to errantly read PID files of another host when its hostname is a prefix of the other hostname.