Add the "prompt" argument to etce-test:
--prompt [ KEYFILE ]

SSH authentication is revised. etce-test first
tries to authenticate through a running ssh-agent.
If that fails, etce-test tries to authenticate
by selecting one of the user's SSH key files directly.
For agent authentication, the correct key must be
the first in the agent's cache. Use the prompt
argument to bypass agent authentication. If the
optional KEYFILE argument is provided, it names the
SSH key to use. This must
either be an absolute file name or the name of
a keyfile in ~/.ssh. If an argument is not provided,
etce-test tries to determine the appropriate key
to use by inspecting ~/.ssh/config.