Add epochtimestamp argument to mgen wrapper (default off).
Fix bug that causes wrappers to store their module relative name, rather than
their fully scoped name, when loaded from a step file with "using"
directives. The bug manifests when a wrapper scoped argument is
specified in a config file it is not passed to the wrapper because of
name mismatch; the config file uses the fully scoped name.